Added reserved custom paths

+++ application/Espo/Core/Utils/File/Manager.php
@@ -51,6 +51,12 @@ class Manager
     protected const GET_SAFE_CONTENTS_RETRY_NUMBER = 10;
     protected const GET_SAFE_CONTENTS_RETRY_INTERVAL = 0.1;
 
+    private const RESERVED_CUSTOM_PATH_LIST = [
+        'custom/Espo/Custom',
+        'custom/Espo/Modules',
+        'client/custom/modules',
+    ];
+
     /**
      * @param ?array{
      *   dir: string|int|null,
@@ -701,6 +707,10 @@ class Manager
         $result = true;
 
         foreach ($dirPaths as $dirPath) {
+            if ($this->isReservedCustomPath($dirPath)) {
+                continue;
+            }
+
             if (is_dir($dirPath) && is_writable($dirPath)) {
                 $result &= rmdir($dirPath);
             }
@@ -1159,5 +1169,20 @@ class Manager
         try {
             opcache_invalidate($filepath, $force);
         } catch (Throwable) {}
+    }
+
+    private function isReservedCustomPath(string $path): bool
+    {
+        $normalizedPath = trim($path, '/');
+
+        foreach (self::RESERVED_CUSTOM_PATH_LIST as $ignoredPath) {
+            $normalizedIgnoredPath = trim($ignoredPath, '/');
+
+            if ($normalizedPath === $normalizedIgnoredPath) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }
